First wave of the new easy-admin.sh

+++ 490-net-ssh/maintainer-ssh-openssh.sh
@@ -10,16 +10,16 @@
 #   BUILDROOT
 #   CHROOT_DIR
 #   INSTANCE
-#   NAMED_CONF
-#   VAR_LIB_NAMED_DIRNAME - useful for multi-instances of 'named' daemons
+#   SSHD_CONF
+#   VAR_LIB_SSHD_DIRNAME - useful for multi-instances of 'sshd' daemons
 
 CHROOT_DIR="${CHROOT_DIR:-}"
 BUILD_DIRNAME="build"
-BUILDROOT="${BUILDROOT:-${BUILD_DIRNAME}/}"
+BUILDROOT="${BUILDROOT:-${BUILD_DIRNAME}}"
 BUILD_SSH="${BUILDROOT}/partial-ssh"
 MINI_REPO="."
 
-source ../easy-admin-installer.sh
+source ./easy-admin-installer.sh
 
 DEFAULT_ETC_CONF_DIRNAME="ssh"
 
@@ -28,44 +28,54 @@ source ../distro-os.sh
 # OpenSSH maintainer settings
 extended_sysconfdir="${sysconfdir}/${DEFAULT_ETC_CONF_DIRNAME}"
 
-sshd_config_filename="sshd_config"
-sshd_config_filespec="${extended_sysconfdir}/$sshd_config_filename"
-ssh_config_filename="ssh_config"
-ssh_config_filespec="${extended_sysconfdir}/$ssh_config_filename"
+SSHD_CONFIG_FILENAME="sshd_config"
+SSHD_CONFIG_FILESPEC="${extended_sysconfdir}/$SSHD_CONFIG_FILENAME"
+SSH_CONFIG_FILENAME="ssh_config"
+SSH_CONFIG_FILESPEC="${extended_sysconfdir}/$SSH_CONFIG_FILENAME"
 
-sshd_configd_dirname="sshd_config.d"
-sshd_configd_dirspec="${extended_sysconfdir}/$sshd_configd_dirname"
-ssh_configd_dirname="ssh_config.d"
-ssh_configd_dirspec="${extended_sysconfdir}/$ssh_configd_dirname"
+SSHD_CONFIGD_DIRNAME="sshd_config.d"
+SSHD_CONFIGD_DIRSPEC="${extended_sysconfdir}/$SSHD_CONFIGD_DIRNAME"
+SSH_CONFIGD_DIRNAME="ssh_config.d"
+SSH_CONFIGD_DIRSPEC="${extended_sysconfdir}/$SSH_CONFIGD_DIRNAME"
 
 # Build area
-
 if [ "${BUILDROOT:0:1}" != '/' ]; then
-  # relative build, create subdirectories
-  # absolute build, do not create build directory
+  # relative build, create all subdirectories
+  # absolute build, do not create build or certain rootfs subdirectories
   if [ ! -d "$BUILDROOT" ]; then
-    mkdir -p "$BUILDROOT"
+    mkdir -p "$BUILDROOT"   # the ONLY '-p' option for mkdir, anywhere
+    # must account for all creations of subdir without using '-p' anywhere else
   fi
-  mkdir -p "${BUILDROOT}$sysconfdir"
+  if [ -n "$CHROOT_DIR" ]; then
+    flex_ckdir "$CHROOT_DIR"
+  fi
+  flex_ckdir "${CHROOT_DIR}/$sysconfdir"
+  flex_ckdir "${CHROOT_DIR}/$extended_sysconfdir"
+  flex_ckdir "${CHROOT_DIR}/$VAR_DIRSPEC"
+  flex_ckdir "${CHROOT_DIR}/$VAR_LIB_DIRSPEC"
+  flex_ckdir "${CHROOT_DIR}/$VAR_LIB_SSH_DIRSPEC"
   BUILD_ABSOLUTE=0
 else
   BUILD_ABSOLUTE=1
 fi
 
+SSH_CONFIG_FILENAME="ssh_config"
+SSHD_CONFIG_FILENAME="sshd_config"
+OPENSSH_CONFIG_DIRSPEC="$extended_sysconfdir"
+SSH_CONFIG_FILESPEC="${OPENSSH_CONFIG_DIRSPEC}/$SSH_CONFIG_FILENAME"
+SSHD_CONFIG_FILESPEC="${OPENSSH_CONFIG_DIRSPEC}/$SSHD_CONFIG_FILENAME"
 
-ssh_config_filename="ssh_config"
-sshd_config_filename="sshd_config"
-openssh_config_dirspec="$extended_sysconfdir"
-ssh_config_filespec="${openssh_config_dirspec}/$ssh_config_filename"
-sshd_config_filespec="${openssh_config_dirspec}/$sshd_config_filename"
+# ssh_confd_dirspec="$extended_sysconfdir/ssh_config.d"
+# sshd_confd_dirspec="$extended_sysconfdir/sshd_config.d"
 
-ssh_confd_dirspec="$extended_sysconfdir/ssh_config.d"
-sshd_confd_dirspec="$extended_sysconfdir/sshd_config.d"
+# FILE_SETTINGS_FILESPEC="${BUILDROOT}/file-settings-ssh-common.sh"
+if [ -n "$FILE_SETTINGS_FILESPEC" ]; then
+  if [ -f "$FILE_SETTINGS_FILESPEC" ]; then
+    rm -f "$FILE_SETTINGS_FILESPEC"
+  fi
+fi
 
-FILE_SETTINGS_FILESPEC="${BUILDROOT}/file-settings-ssh-common.sh"
-rm -f "$FILE_SETTINGS_FILESPEC"
-
-echo "Detected $ID distro."
+[[ -n "$VERBOSE" ]] && echo "Detected $ID distro."
 
 case $ID in
   debian|devuan)
@@ -188,20 +198,69 @@ case $ID in
     ;;
 esac
 
-flex_mkdir "$sysconfdir"
+flex_ckdir "$sysconfdir"
+
+
+if [ 0 -ne 0 ]; then
+# We will assume that /etc/ssh is not installed, so create them anyway
 flex_mkdir "$extended_sysconfdir"
+
 if [ "$HAS_SSHD_CONFIG_D" -ne 0 ]; then
-  flex_mkdir "$sshd_configd_dirspec"
+  flex_mkdir "$SSHD_CONFIGD_DIRSPEC"
 fi
+exit
 
 if [ "$HAS_SSH_CONFIG_D" -ne 0 ]; then
-  flex_mkdir "$ssh_configd_dirspec"
+  flex_mkdir "$SSH_CONFIGD_DIRSPEC"
+fi
+fi  # [0 -ne 0] needs copying into other main SSH scripts
+
+SSH_FOUND=0
+EXECUTABLE_SSH=0
+SYNTAX_CHECKABLE_SSH=0
+SUDO_REQUIRED_SSH=0
+SSH_BIN_FILESPEC="$( whereis ssh | awk '{print $2}')"
+if [ -z "$SSH_BIN_FILESPEC" ]; then
+  echo "Binary $SSH_BIN_FILESPEC not found in \$PATH"
+  echo "No syntax checking will be done."
+else
+  SSH_FOUND=1
+fi
+if [ ! -f "$SSH_BIN_FILESPEC" ]; then
+  echo "Binary $SSH_BIN_FILESPEC does not exist."
+  echo "'sudo' may be required to do optional syntax-checking"
+else
+  SYNTAX_CHECKABLE_SSH=1
+fi
+SSH_BIN_GROUP="$(stat -c %G "$SSH_BIN_FILESPEC")"
+if [ -x "$SSH_BIN_FILESPEC" ]; then
+  EXECUTABLE_SSH=1
+else
+  SUDO_REQUIRED_SSH=1
 fi
 
-ssh_bin_filespec="$( whereis ssh | awk '{print $2}')"
-sshd_bin_filespec="$( whereis sshd | awk '{print $2}')"
+SSHD_FOUND=0
+EXECUTABLE_SSHD=0
+SYNTAX_CHECKABLE_SSHD=0
+SUDO_REQUIRED_SSHD=0
+SSHD_BIN_FILESPEC="$( whereis sshd | awk '{print $2}')"
+if [ -z "$SSHD_BIN_FILESPEC" ]; then
+  echo "Binary $SSHD_BIN_FILESPEC not found in \$PATH"
+  echo "No syntax checking will be done."
+else
+  SSHD_FOUND=1
+fi
+if [ ! -f "$SSHD_BIN_FILESPEC" ]; then
+  echo "Binary $SSHD_BIN_FILESPEC does not exist."
+  echo "'sudo' may be required to do optional syntax-checking"
+else
+  SYNTAX_CHECKABLE_SSHD=1
+fi
+SSHD_BIN_GROUP="$(stat -c %G "$SSHD_BIN_FILESPEC")"
+if [ -x "$SSHD_BIN_FILESPEC" ]; then
+  EXECUTABLE_SSHD=1
+else
+  SUDO_REQUIRED_SSHD=1
+fi
 
-
-sshd_home_dirspec="$( egrep "^${SSHD_USER_NAME}:" /etc/passwd | awk -F: '{print $6 }')"
-echo "sshd_home_dirspec: $sshd_home_dirspec"
-
+SSHD_HOME_DIRSPEC="$( egrep "^${SSHD_USER_NAME}:" /etc/passwd | awk -F: '{print $6 }')"
